debug DX12 SYCL interop image copy

+++ src/Graphics/D3D12/Utils/Device.hpp
@@ -89,7 +89,7 @@ private:
     // Device information (retrieved from adapter).
     std::string adapterName;
     uint32_t vendorId;
-    uint32_t adapterLuid;
+    uint64_t adapterLuid;
 };
 
 }}
